db: moved migrations inside db mod
