+++ store/src/lib.rs
@@ -6,8 +6,6 @@ pub mod server;
 pub mod state;
 pub mod types;
 
-mod migrations;
-
 // CONSTANTS
 // =================================================================================================
 pub const COMPONENT: &str = "miden-store";
